refactor: MusicStreaming으로 이름 변경(#19)
