+++ src/main/java/play/pluv/oauth/spotify/SpotifyPlayListResponses.java
@@ -1,6 +1,6 @@
 package play.pluv.oauth.spotify;
 
-import static play.pluv.playlist.domain.PlayListProvider.SPOTIFY;
+import static play.pluv.music.domain.MusicStreaming.SPOTIFY;
 
 import java.util.List;
 import play.pluv.playlist.domain.PlayList;
