+++ src/main/java/play/pluv/playlist/application/MusicPlatformConnector.java
@@ -2,12 +2,11 @@ package play.pluv.playlist.application;
 
 import java.util.List;
 import play.pluv.playlist.domain.PlayList;
-import play.pluv.playlist.domain.PlayListProvider;
+import play.pluv.music.domain.MusicStreaming;
 
 public interface MusicPlatformConnector {
 
-  //TODO : 최신순 정렬 -> 불가능
   List<PlayList> getPlayList(final String authCode);
 
-  PlayListProvider supportedType();
+  MusicStreaming supportedType();
 }
