+++ src/main/java/play/pluv/playlist/domain/PlayListId.java
@@ -1,8 +1,10 @@
 package play.pluv.playlist.domain;
 
+import play.pluv.music.domain.MusicStreaming;
+
 public record PlayListId(
     String id,
-    PlayListProvider playListProvider
+    MusicStreaming musicStreaming
 ) {
 
 }
