+++ src/test/java/play/pluv/playlist/application/PlayListServiceTest.java
@@ -3,7 +3,7 @@ package play.pluv.playlist.application;
 import static org.assertj.core.api.Assertions.assertThat;
 import static play.pluv.fixture.PlayListFixture.스포티파이_플레이리스트_1;
 import static play.pluv.fixture.PlayListFixture.스포티파이_플레이리스트_2;
-import static play.pluv.playlist.domain.PlayListProvider.SPOTIFY;
+import static play.pluv.music.domain.MusicStreaming.SPOTIFY;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
